refactor: getLoggedIn을 services에서 repository 세그먼트로 수정
